cargo: Use rye-vendored Mocktopus distribution
I have not seen or heard any activity from the maintainer of Mocktopus
in the last month, and there are a number of breaking issues on the
nightly end.  This commit should make the code compile and test nicely
again after a feature stabilized in the nightly that broke Mocktopus'
code (get_test_id -> test_id with no deprecation), and moves us away
from the upstream dependency.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>
Tested-by: Kristofer Rye <kristofer.rye@gmail.com>